+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes
 
+## unreleased
+
+- Added command to write selected speaker areas from a dataset to GeoJSON, optionally including
+  corresponding Glottolog point coordinates.
+
+
 ## [1.3.0] - 2025-01-28
 
 - Added command to compute multi-polygon spread for speaker areas of language-level items in the